feat: 結果リストを results 窓へ移設(snapshot 一方向 + main driver)(#646 PR2・決定 5/6/7)
- ResultsView が RowsSnapshot を描き、クリックは共有スロットで main へ逆流
- main は bar(+toast)固定高になり、results の可視性/サイズ/位置を駆動
- 旧 compute_window_height/HeightParams を撤去(実件数フィットへ全面移行)
